refactor(orchestrator): D38 split builtin_tools by tool family
Extract agent-gate, workspace FS/shell, and git/web/media handlers into
dedicated modules; keep builtin_tools as registry/dispatch facade with
re-exports. Retarget bg_job monkeypatch to workspace module.
