+++ memory/DELIVERABLES.md
@@ -4,6 +4,15 @@
 > **何时写：** 每个**代码 Task** 完成并 commit 后，在 Check-out 追加一节（见 `CLAUDE.md` §Check-out）。  
 > **生命周期：** 见 [`docs/document-governance.md`](../docs/document-governance.md) §文档生命周期；冷数据见 [`archive/`](./archive/)。
 
+## D38 P2 · builtin_tools 工具族拆分（2026-08-01）
+
+| 项 | 说明 |
+|----|------|
+| **Branch / PR** | `cursor/d38-chat-runner-split-9d3c` · [#98](https://github.com/fancy1108/Clutch/pull/98) |
+| **Changed** | `builtin_tools_{agent,workspace,external}.py` + facade re-export |
+| **Verification** | pytest **942 passed / 8 skipped** |
+| **Stat** | `builtin_tools.py` ≈2228→≈751 |
+
 ## D38 Phase 2 · chat_runner plain/workflow（2026-08-01）
 
 | 项 | 说明 |
