+++ services/orchestrator/src/builtin_tools.py
@@ -13,12 +13,77 @@ from typing import Any
 
 CLUTCH_TOOLS_SERVER_ID = "clutch-tools"
 
-_MAX_READ_CHARS = 120_000
-_MAX_GREP_HITS = 50
-_MAX_LIST_ENTRIES = 200
-_DEFAULT_CMD_TIMEOUT_S = 60
-_MAX_CMD_OUTPUT_CHARS = 80_000
 
+# D38 — tool-family modules (re-exported for routes/tests)
+from src.builtin_tools_agent import (
+    _DIFF_FILE_STATUSES,
+    _STEP_INDEX_RE,
+    _TODO_STATUSES,
+    _VERIFICATION_CONCLUSIONS,
+    _VERIFICATION_STEP_STATUSES,
+    _basename_path,
+    _coerce_todos_list,
+    _git_diff_for_path,
+    _hunk_from_old_new,
+    _parse_unified_diff_lines,
+    _tool_ask_user_question,
+    _tool_delegate_subtask,
+    _tool_diagnostics,
+    _tool_goal_write,
+    _tool_propose_plan,
+    _tool_read_skill,
+    _tool_remember_preference,
+    _tool_submit_diff_summary,
+    _tool_submit_verification,
+    _tool_todo_write,
+    _truncate_patch,
+    build_diff_summary_from_paths,
+    build_inline_edit_diff_cards,
+    enrich_diff_file_entry,
+    is_ask_user_question_tool,
+    is_delegate_subtask_tool,
+    is_goal_write_tool,
+    is_propose_plan_tool,
+    is_read_skill_tool,
+    is_submit_diff_summary_tool,
+    is_submit_verification_tool,
+    is_todo_write_tool,
+    normalize_diff_summary,
+    normalize_goal_args,
+    normalize_plan_args,
+    normalize_question_args,
+    normalize_todo_items,
+    normalize_verification_report,
+    parse_question_selection,
+    strip_plan_step_index,
+)
+from src.builtin_tools_workspace import (
+    _DEFAULT_CMD_TIMEOUT_S,
+    _MAX_CMD_OUTPUT_CHARS,
+    _MAX_DIFF_PATCH_LINES,
+    _MAX_GREP_HITS,
+    _MAX_LIST_ENTRIES,
+    _MAX_READ_CHARS,
+    _bg_job_run_id,
+    _tool_apply_patch,
+    _tool_grep,
+    _tool_kill_background_job,
+    _tool_list_background_jobs,
+    _tool_list_dir,
+    _tool_read_file,
+    _tool_run_terminal_cmd,
+    _tool_search_replace,
+)
+from src.builtin_tools_external import (
+    _run_git,
+    _tool_generate_image,
+    _tool_generate_video,
+    _tool_git_commit,
+    _tool_git_diff,
+    _tool_git_status,
+    _tool_web_fetch,
+    _tool_web_search,
+)
 
 def resolve_clutch_tools_server() -> dict[str, Any] | None:
     from src.workspace import get_workspace
@@ -640,710 +705,6 @@ def list_builtin_tools() -> list[dict[str, Any]]:
     return tools
 
 
-def is_propose_plan_tool(name: str) -> bool:
-    short = name.split("__")[-1].lower().replace("-", "_")
-    return short == "propose_plan"
-
-
-def is_todo_write_tool(name: str) -> bool:
-    short = name.split("__")[-1].lower().replace("-", "_")
-    return short in {"todo_write", "write_todos", "update_todos"}
-
-
-def is_goal_write_tool(name: str) -> bool:
-    short = name.split("__")[-1].lower().replace("-", "_")
-    return short in {"goal_write", "set_goal", "update_goal"}
-
-
-def is_ask_user_question_tool(name: str) -> bool:
-    short = name.split("__")[-1].lower().replace("-", "_")
-    return short in {"ask_user_question", "ask_question", "user_question"}
-
-
-def is_submit_verification_tool(name: str) -> bool:
-    short = name.split("__")[-1].lower().replace("-", "_")
-    return short in {
-        "submit_verification",
-        "verification_report",
-        "submit_verification_report",
-    }
-
-
-def is_submit_diff_summary_tool(name: str) -> bool:
-    short = name.split("__")[-1].lower().replace("-", "_")
-    return short in {
-        "submit_diff_summary",
-        "diff_summary",
-        "propose_diff_review",
-    }
-
-
-def is_read_skill_tool(name: str) -> bool:
-    short = name.split("__")[-1].lower().replace("-", "_")
-    return short in {"read_skill", "load_skill"}
-
-
-def is_delegate_subtask_tool(name: str) -> bool:
-    short = name.split("__")[-1].lower().replace("-", "_")
-    return short == "delegate_subtask"
-
-
-_TODO_STATUSES = frozenset({"pending", "in_progress", "completed"})
-_VERIFICATION_STEP_STATUSES = frozenset({"passed", "failed", "skipped"})
-_VERIFICATION_CONCLUSIONS = frozenset({"passed", "failed"})
-_DIFF_FILE_STATUSES = frozenset({"A", "M", "D"})
-_MAX_DIFF_PATCH_LINES = 160
-
-
-def normalize_goal_args(func_args: dict[str, Any] | None) -> dict[str, Any]:
-    payload = func_args if isinstance(func_args, dict) else {}
-    title = str(payload.get("title") or payload.get("goal") or "").strip()
-    if not title:
-        title = "Goal"
-    try:
-        progress = int(payload.get("progress") if payload.get("progress") is not None else 0)
-    except (TypeError, ValueError):
-        progress = 0
-    progress = max(0, min(100, progress))
-    done = bool(payload.get("done"))
-    if progress >= 100:
-        done = True
-        progress = 100
-    return {"title": title, "progress": progress, "done": done}
-
-
-def _coerce_todos_list(value: Any) -> list[Any]:
-    """Coerce tool `todos` payload to a list (models often JSON-stringify the array)."""
-    if value is None:
-        return []
-    if isinstance(value, list):
-        return value
-    if isinstance(value, dict):
-        # Single todo object, or nested {"todos": ...}
-        if "todos" in value:
-            return _coerce_todos_list(value.get("todos"))
-        if any(key in value for key in ("content", "text", "title", "status", "id")):
-            return [value]
-        return []
-    if isinstance(value, str):
-        text = value.strip()
-        if not text:
-            return []
-        try:
-            parsed = json.loads(text)
-        except json.JSONDecodeError:
-            # Do not list(str) — that explodes into one todo per character (D8 PM).
-            return []
-        return _coerce_todos_list(parsed)
-    return []
-
-
-def normalize_todo_items(
-    raw: Any,
-    *,
-    existing: list[dict[str, Any]] | None = None,
-    merge: bool = False,
-) -> list[dict[str, Any]]:
-    """Normalize todo_write args into [{id, content, status}, ...]."""
-    items_in: list[Any]
-    if isinstance(raw, dict) and "todos" in raw:
-        merge = bool(raw.get("merge")) or merge
-        items_in = _coerce_todos_list(raw.get("todos"))
-    elif isinstance(raw, list):
-        items_in = raw
-    elif isinstance(raw, str):
-        items_in = _coerce_todos_list(raw)
-    else:
-        items_in = []
-
-    normalized: list[dict[str, Any]] = []
-    for idx, item in enumerate(items_in):
-        if isinstance(item, str):
-            content = item.strip()
-            status = "pending"
-            todo_id = f"todo_{idx + 1}"
-        elif isinstance(item, dict):
-            content = str(item.get("content") or item.get("text") or item.get("title") or "").strip()
-            status = str(item.get("status") or "pending").strip().lower().replace("-", "_")
-            if status == "done":
-                status = "completed"
-            if status == "doing" or status == "active":
-                status = "in_progress"
-            if status not in _TODO_STATUSES:
-                status = "pending"
-            todo_id = str(item.get("id") or f"todo_{idx + 1}").strip() or f"todo_{idx + 1}"
-        else:
-            continue
-        if not content:
-            continue
-        normalized.append({"id": todo_id, "content": content, "status": status})
-
-    if not merge or not existing:
-        return normalized
-
-    by_id = {str(t.get("id")): dict(t) for t in existing if isinstance(t, dict)}
-    order = [str(t.get("id")) for t in existing if isinstance(t, dict)]
-    for item in normalized:
-        tid = str(item["id"])
-        if tid in by_id:
-            by_id[tid] = item
-        else:
-            by_id[tid] = item
-            order.append(tid)
-    return [by_id[tid] for tid in order if tid in by_id]
-
-
-_STEP_INDEX_RE = re.compile(r"^\s*(?:\d+[\.\)\:．、]\s*|\d+\s+)")
-
-
-def strip_plan_step_index(text: str) -> str:
-    """Remove leading '1.' / '1)' (repeat) so PlanCard does not show '1. 1. …'."""
-    cleaned = (text or "").strip()
-    for _ in range(4):
-        next_text = _STEP_INDEX_RE.sub("", cleaned).strip()
-        if next_text == cleaned:
-            break
-        cleaned = next_text
-    return cleaned or (text or "").strip()
-
-
-def normalize_plan_args(func_args: dict[str, Any] | None) -> dict[str, Any]:
-    payload = func_args if isinstance(func_args, dict) else {}
-    title = str(payload.get("title") or "Plan").strip() or "Plan"
-    raw_steps = payload.get("steps")
-    steps: list[str] = []
-    if isinstance(raw_steps, list):
-        for item in raw_steps:
-            text = strip_plan_step_index(str(item))
-            if text:
-                steps.append(text)
-    elif isinstance(raw_steps, str) and raw_steps.strip():
-        steps = [
-            strip_plan_step_index(line.strip(" -*\t"))
-            for line in raw_steps.splitlines()
-            if line.strip()
-        ]
-        steps = [s for s in steps if s]
-    summary = str(payload.get("summary") or "").strip()
-    return {"title": title, "steps": steps, "summary": summary}
-
-
-def normalize_question_args(func_args: dict[str, Any] | None) -> dict[str, Any]:
-    payload = func_args if isinstance(func_args, dict) else {}
-    question = str(payload.get("question") or payload.get("prompt") or "").strip()
-    raw_opts = payload.get("options")
-    options: list[dict[str, str]] = []
-    if isinstance(raw_opts, list):
-        for idx, item in enumerate(raw_opts):
-            if isinstance(item, str):
-                label = item.strip()
-                if not label:
-                    continue
-                options.append({"id": f"opt_{idx + 1}", "label": label})
-            elif isinstance(item, dict):
-                label = str(item.get("label") or item.get("text") or item.get("title") or "").strip()
-                if not label:
-                    continue
-                oid = str(item.get("id") or f"opt_{idx + 1}").strip() or f"opt_{idx + 1}"
-                options.append({"id": oid, "label": label})
-    allow_custom = payload.get("allow_custom")
-    if allow_custom is None:
-        allow_custom = True
-    return {
-        "question": question or "Please choose an option",
-        "options": options,
-        "allow_custom": bool(allow_custom),
-    }
-
-
-def parse_question_selection(
-    instructions: str,
-    func_args: dict[str, Any] | None,
-) -> dict[str, str]:
-    """Map human_decision instructions to {id, label}."""
-    normalized = normalize_question_args(func_args)
-    options = normalized["options"]
-    raw = (instructions or "").strip()
-    if raw:
-        try:
-            parsed = json.loads(raw)
-            if isinstance(parsed, dict):
-                label = str(parsed.get("label") or parsed.get("text") or "").strip()
-                oid = str(parsed.get("id") or "").strip()
-                if label or oid:
-                    if not label and oid:
-                        for opt in options:
-                            if opt["id"] == oid:
-                                label = opt["label"]
-                                break
-                        label = label or oid
-                    if not oid:
-                        oid = next((o["id"] for o in options if o["label"] == label), f"custom_{label[:24]}")
-                    return {"id": oid, "label": label}
-        except json.JSONDecodeError:
-            pass
-        for opt in options:
-            if opt["id"] == raw or opt["label"] == raw:
-                return dict(opt)
-        return {"id": "custom", "label": raw}
-    if options:
-        return dict(options[0])
-    return {"id": "custom", "label": "(no answer)"}
-
-
-def normalize_verification_report(
-    func_args: dict[str, Any] | None,
-    *,
-    existing_todos: list[dict[str, Any]] | None = None,
-) -> dict[str, Any]:
-    """Normalize submit_verification args; force failed if todos incomplete."""
-    payload = func_args if isinstance(func_args, dict) else {}
-    title = str(payload.get("title") or payload.get("name") or "").strip() or "Verification"
-    raw_conclusion = str(payload.get("conclusion") or payload.get("status") or "failed").strip().lower()
-    conclusion = raw_conclusion if raw_conclusion in _VERIFICATION_CONCLUSIONS else "failed"
-    steps: list[dict[str, str]] = []
-    raw_steps = payload.get("steps")
-    if isinstance(raw_steps, list):
-        for idx, item in enumerate(raw_steps):
-            if isinstance(item, str):
-                name = item.strip()
-                if name:
-                    steps.append(
-                        {
-                            "id": f"step_{idx + 1}",
-                            "name": name,
-                            "status": "passed" if conclusion == "passed" else "failed",
-                            "detail": "",
-                        }
-                    )
-                continue
-            if not isinstance(item, dict):
-                continue
-            name = str(item.get("name") or item.get("title") or item.get("check") or "").strip()
-            if not name:
-                continue
-            status = str(item.get("status") or "failed").strip().lower()
-            if status not in _VERIFICATION_STEP_STATUSES:
-                status = "failed"
-            detail = str(item.get("detail") or item.get("message") or "").strip()
-            sid = str(item.get("id") or f"step_{idx + 1}").strip() or f"step_{idx + 1}"
-            steps.append({"id": sid, "name": name, "status": status, "detail": detail})
-    summary = str(payload.get("summary") or "").strip()
-    next_actions: list[str] = []
-    raw_next = payload.get("next_actions") or payload.get("nextActions")
-    if isinstance(raw_next, list):
-        for item in raw_next:
-            text = str(item).strip()
-            if text:
-                next_actions.append(text)
-    changed_files: list[str] = []
-    raw_files = payload.get("changed_files") or payload.get("changedFiles")
-    if isinstance(raw_files, list):
-        for item in raw_files:
-            path = str(item).strip()
-            if path and path not in changed_files:
-                changed_files.append(path)
-
-    incomplete = [
-        t
-        for t in (existing_todos or [])
-        if isinstance(t, dict) and str(t.get("status") or "") != "completed"
-    ]
-    if conclusion == "passed" and incomplete:
-        conclusion = "failed"
-        names = ", ".join(
-            str(t.get("content") or t.get("id") or "todo").strip() for t in incomplete[:5]
-        )
-        steps.insert(
-            0,
-            {
-                "id": "todos_incomplete",
-                "name": "Session todos incomplete",
-                "status": "failed",
-                "detail": f"Cannot claim passed while todos remain open: {names}",
-            },
-        )
-        tip = "Mark remaining todos completed (or cancel them) before claiming success."
-        if tip not in next_actions:
-            next_actions.insert(0, tip)
-        if not summary:
-            summary = "Verification forced to failed because session todos are still incomplete."
-
-    if any(s["status"] == "failed" for s in steps) and conclusion == "passed":
-        conclusion = "failed"
-
-    return {
-        "title": title,
-        "conclusion": conclusion,
-        "steps": steps,
-        "summary": summary,
-        "nextActions": next_actions,
-        "changedFiles": changed_files,
-    }
-
-
-def _truncate_patch(patch: str, *, max_lines: int = _MAX_DIFF_PATCH_LINES) -> str:
-    lines = patch.splitlines()
-    if len(lines) <= max_lines:
-        return patch.strip()
-    kept = lines[:max_lines]
-    kept.append(f"... ({len(lines) - max_lines} more lines truncated)")
-    return "\n".join(kept).strip()
-
-
-def _parse_unified_diff_lines(patch: str) -> list[dict[str, Any]]:
-    """Map unified diff text → DiffLine-shaped dicts for Chat/Changes."""
-    out: list[dict[str, Any]] = []
-    line_num = 0
-    for raw in patch.splitlines():
-        if raw.startswith("+++") or raw.startswith("---") or raw.startswith("diff ") or raw.startswith("index "):
-            continue
-        if raw.startswith("@@"):
-            # @@ -a,b +c,d @@ → start from new-file line c
-            m = re.search(r"\+(\d+)", raw)
-            if m:
-                line_num = max(0, int(m.group(1)) - 1)
-            continue
-        if raw.startswith("+"):
-            line_num += 1
-            out.append({"lineNum": line_num, "type": "addition", "text": raw[1:]})
-        elif raw.startswith("-"):
-            out.append({"lineNum": line_num, "type": "deletion", "text": raw[1:]})
-        elif raw.startswith("\\"):
-            continue
-        else:
-            text = raw[1:] if raw.startswith(" ") else raw
-            line_num += 1
-            out.append({"lineNum": line_num, "type": "normal", "text": text})
-    return out
-
-
-def _git_diff_for_path(rel: str) -> tuple[str, str]:
-    """Return (status A|M|D, unified patch) for a workspace-relative path."""
-    from src.workspace import WorkspaceError, require_workspace, resolve_allowed_path
-
-    try:
-        root = require_workspace()
-        target = resolve_allowed_path(rel)
-    except WorkspaceError:
-        return "M", ""
-
-    rel_posix = rel.replace("\\", "/")
-    try:
-        porcelain = subprocess.run(
-            ["git", "status", "--porcelain", "--", rel_posix],
-            cwd=str(root),
-            capture_output=True,
-            text=True,
-            timeout=8,
-            check=False,
-        )
-    except (OSError, subprocess.TimeoutExpired):
-        porcelain = None
-
-    status = "M"
-    if porcelain and porcelain.returncode == 0 and porcelain.stdout.strip():
-        code = porcelain.stdout.strip()[:2]
-        if "A" in code or "?" in code:
-            status = "A"
-        elif "D" in code:
-            status = "D"
-        else:
-            status = "M"
-
-    patch = ""
-    try:
-        if status == "A" or (porcelain and porcelain.stdout.strip().startswith("??")):
-            diff = subprocess.run(
-                ["git", "diff", "--no-index", "--", "/dev/null", rel_posix],
-                cwd=str(root),
-                capture_output=True,
-                text=True,
-                timeout=8,
-                check=False,
-            )
-            # git --no-index exits 1 when files differ
-            patch = diff.stdout or ""
-        else:
-            diff = subprocess.run(
-                ["git", "diff", "HEAD", "--", rel_posix],
-                cwd=str(root),
-                capture_output=True,
-                text=True,
-                timeout=8,
-                check=False,
-            )
-            patch = diff.stdout or ""
-            if not patch.strip():
-                diff = subprocess.run(
-                    ["git", "diff", "--", rel_posix],
-                    cwd=str(root),
-                    capture_output=True,
-                    text=True,
-                    timeout=8,
-                    check=False,
-                )
-                patch = diff.stdout or ""
-    except (OSError, subprocess.TimeoutExpired):
-        patch = ""
-
-    if not patch.strip() and target.is_file() and status != "D":
-        try:
-            text = target.read_text(encoding="utf-8", errors="replace")
-        except OSError:
-            text = ""
-        preview = text.splitlines()[:80]
-        patch = "\n".join(f"+{line}" for line in preview)
-        if len(text.splitlines()) > 80:
-            patch += f"\n... ({len(text.splitlines()) - 80} more lines truncated)"
-        status = status if status in _DIFF_FILE_STATUSES else "A"
-
-    return status, _truncate_patch(patch)
-
-
-def enrich_diff_file_entry(entry: dict[str, Any]) -> dict[str, Any]:
-    """Fill status/patch/diffs for one file entry when missing."""
-    path = str(entry.get("path") or "").strip()
-    if not path:
-        return entry
-    status = str(entry.get("status") or "").strip().upper()[:1]
-    if status not in _DIFF_FILE_STATUSES:
-        status = ""
-    patch = str(entry.get("patch") or "").strip()
-    if not patch or not status:
-        git_status, git_patch = _git_diff_for_path(path)
-        if not status:
-            status = git_status
-        if not patch:
-            patch = git_patch
-    patch = _truncate_patch(patch) if patch else ""
-    diffs = entry.get("diffs")
-    if not isinstance(diffs, list) or not diffs:
-        diffs = _parse_unified_diff_lines(patch) if patch else []
-    summary = str(entry.get("summary") or "").strip()
-    return {
-        "path": path,
-        "status": status or "M",
-        "summary": summary,
-        "patch": patch,
-        "diffs": diffs,
-    }
-
-
-def normalize_diff_summary(
-    func_args: dict[str, Any] | None,
-    *,
-    enrich: bool = True,
-) -> dict[str, Any]:
-    """Normalize submit_diff_summary args into a DiffSummary card payload."""
-    payload = func_args if isinstance(func_args, dict) else {}
-    title = str(payload.get("title") or payload.get("name") or "").strip() or "Changes"
-    summary = str(payload.get("summary") or "").strip()
-    files_in: list[Any] = []
-    raw_files = payload.get("files")
-    if isinstance(raw_files, list):
-        files_in = raw_files
-    elif isinstance(payload.get("changed_files") or payload.get("changedFiles"), list):
-        files_in = list(payload.get("changed_files") or payload.get("changedFiles") or [])
-
-    files: list[dict[str, Any]] = []
-    seen: set[str] = set()
-    for item in files_in:
-        if isinstance(item, str):
-            path = item.strip()
-            if not path or path in seen:
-                continue
-            seen.add(path)
-            entry: dict[str, Any] = {"path": path}
-        elif isinstance(item, dict):
-            path = str(item.get("path") or item.get("file") or item.get("name") or "").strip()
-            if not path or path in seen:
-                continue
-            seen.add(path)
-            entry = dict(item)
-            entry["path"] = path
-        else:
-            continue
-        if enrich:
-            files.append(enrich_diff_file_entry(entry))
-        else:
-            path = entry["path"]
-            status = str(entry.get("status") or "M").strip().upper()[:1]
-            if status not in _DIFF_FILE_STATUSES:
-                status = "M"
-            patch = _truncate_patch(str(entry.get("patch") or ""))
-            diffs = entry.get("diffs")
-            if not isinstance(diffs, list) or not diffs:
-                diffs = _parse_unified_diff_lines(patch) if patch else []
-            files.append(
-                {
-                    "path": path,
-                    "status": status,
-                    "summary": str(entry.get("summary") or "").strip(),
-                    "patch": patch,
-                    "diffs": diffs,
-                }
-            )
-
-    if not summary and files:
-        summary = f"{len(files)} file(s) changed"
-
-    out: dict[str, Any] = {"title": title, "summary": summary, "files": files}
-    if payload.get("inline") is True:
-        out["inline"] = True
-    return out
-
-
-def build_diff_summary_from_paths(
-    paths: list[str] | None,
-    *,
-    title: str = "Changes",
-) -> dict[str, Any] | None:
-    """Auto-build a DiffSummary from workspace-relative paths (D6 seal fallback)."""
-    cleaned: list[str] = []
-    for path in paths or []:
-        rel = str(path).strip()
-        if rel and rel not in cleaned:
-            cleaned.append(rel)
-    if not cleaned:
-        return None
-    return normalize_diff_summary({"title": title, "files": cleaned}, enrich=True)
-
-
-def _basename_path(path: str) -> str:
-    normalized = path.replace("\\", "/")
-    parts = [p for p in normalized.split("/") if p]
-    return parts[-1] if parts else path
-
-
-def _hunk_from_old_new(old_s: str, new_s: str) -> str:
-    """Line-level unified hunk so appends don't re-paint unchanged lines as -/+."""
-    old_lines = str(old_s).splitlines()
-    new_lines = str(new_s).splitlines()
-    if old_lines == new_lines:
-        return ""
-    diff_lines = [
-        line
-        for line in difflib.unified_diff(old_lines, new_lines, lineterm="", n=2)
-        if not line.startswith("---") and not line.startswith("+++")
-    ]
-    if diff_lines:
-        return "\n".join(diff_lines)
-    # Fallback: whole block replace (should be rare).
-    parts: list[str] = ["@@ -1 +1 @@"]
-    parts.extend(f"-{line}" for line in old_lines)
-    parts.extend(f"+{line}" for line in new_lines)
-    return "\n".join(parts)
-
-
-def build_inline_edit_diff_cards(
-    *,
-    tool_name: str,
-    func_args: dict[str, Any],
-    result_str: str,
-) -> list[dict[str, Any]]:
-    """
-    Cursor-style: one DiffSummary card per edited file, right after the edit tool.
-    Prefer the edit payload (old/new or patch) so the chat shows the hunk immediately.
-    """
-    if result_str.startswith("Error executing tool"):
-        return []
-    short = tool_name.split("__")[-1].lower().replace("-", "_")
-    cards: list[dict[str, Any]] = []
-
-    if short == "search_replace":
-        path = str(func_args.get("path") or "").strip()
-        old_s = func_args.get("old_string")
-        new_s = func_args.get("new_string")
-        if path and old_s is not None and new_s is not None:
-            try:
-                payload = json.loads(result_str)
-            except json.JSONDecodeError:
-                payload = None
-            if isinstance(payload, dict) and payload.get("path"):
-                path = str(payload["path"]).strip() or path
-            patch = _hunk_from_old_new(str(old_s), str(new_s))
-            entry = enrich_diff_file_entry(
-                {"path": path, "status": "M", "patch": patch}
-            )
-            cards.append(
-                {
-                    "title": _basename_path(path),
-                    "summary": "",
-                    "files": [entry],
-                    "inline": True,
-                }
-            )
-            return cards
-
-    if short == "apply_patch":
-        patch = str(func_args.get("patch") or "").strip()
-        paths: list[str] = []
-        try:
-            payload = json.loads(result_str)
-        except json.JSONDecodeError:
-            payload = None
-        if isinstance(payload, dict):
-            for raw in payload.get("changed_paths") or []:
-                rel = str(raw).strip()
-                if rel and rel not in paths:
-                    paths.append(rel)
-        if not paths and patch:
-            try:
-                from src.apply_patch import extract_patch_paths
-
-                paths = list(extract_patch_paths(patch) or [])
-            except Exception:
-                paths = []
-        for path in paths:
-            entry = enrich_diff_file_entry({"path": path, "patch": patch if len(paths) == 1 else ""})
-            cards.append(
-                {
-                    "title": _basename_path(path),
-                    "summary": "",
-                    "files": [entry],
-                    "inline": True,
-                }
-            )
-        return cards
-
-    # Generic write/create/edit — show git/content diff for the path.
-    from src.mcp_risk import extract_mcp_file_paths
-
-    paths = extract_mcp_file_paths(tool_name, func_args)
-    if not paths:
-        try:
-            payload = json.loads(result_str)
-        except json.JSONDecodeError:
-            payload = None
-        if isinstance(payload, dict):
-            for raw in payload.get("changed_paths") or []:
-                rel = str(raw).strip()
-                if rel and rel not in paths:
-                    paths.append(rel)
-            if payload.get("path"):
-                rel = str(payload["path"]).strip()
-                if rel and rel not in paths:
-                    paths.append(rel)
-    editish = any(
-        tok in short
-        for tok in ("write", "edit", "create", "patch", "replace", "delete")
-    )
-    if not editish:
-        return []
-    for path in paths:
-        entry = enrich_diff_file_entry({"path": path})
-        cards.append(
-            {
-                "title": _basename_path(path),
-                "summary": "",
-                "files": [entry],
-                "inline": True,
-            }
-        )
-    return cards
-
-
 def execute_builtin_tool(tool_name: str, arguments: dict[str, Any]) -> str:
     handlers = {
         "read_file": _tool_read_file,
@@ -1388,841 +749,3 @@ def execute_builtin_tool(tool_name: str, arguments: dict[str, Any]) -> str:
         return f"Error executing tool: {exc}"
 
 
-def _tool_read_skill(arguments: dict[str, Any]) -> str:
-    from src.agent_skills import load_skill_body
-
-    key = str(arguments.get("key") or arguments.get("skill") or "").strip()
-    if not key:
-        return "Error executing tool: read_skill requires `key` (skill registry key)"
-    body = load_skill_body(key)
-    if not body:
-        return (
-            f"Error executing tool: skill `{key}` not found in the Skills Registry. "
-            "Check the Skills catalog keys bound to this agent."
-        )
-    return body
-
-
-def _tool_delegate_subtask(arguments: dict[str, Any]) -> str:
-    from src.subagent_runner import (
-        default_subtask_max_steps,
-        delegate_result_json,
-        get_delegate_context,
-        run_subagent,
-    )
-
-    try:
-        ctx = get_delegate_context()
-        if not ctx:
-            return (
-                "Error executing tool: delegate_subtask requires an active Chat agent context"
-            )
-        task_type = str(arguments.get("type") or "explore")
-        raw_max = ctx.get("max_steps")
-        steps = (
-            int(raw_max)
-            if raw_max is not None
-            else default_subtask_max_steps(task_type)
-        )
-        card = run_subagent(
-            task_type=task_type,
-            prompt=str(arguments.get("prompt") or ""),
-            title=str(arguments.get("title") or "") or None,
-            servers=list(ctx.get("servers") or []),
-            model_id=ctx.get("model_id"),
-            on_log=ctx.get("on_log"),
-            on_subtask_update=ctx.get("on_subtask_update"),
-            max_steps=steps,
-            permission_mode=str(ctx.get("permission_mode") or "auto_edit"),
-            pause_on_risky=bool(ctx.get("pause_on_risky", True)),
-            subtask_id=ctx.get("subtask_id"),
-        )
-        return delegate_result_json(card)
-    except ValueError as exc:
-        return f"Error executing tool: {exc}"
-    except Exception as exc:
-        return f"Error executing tool: {exc}"
-
-
-def _tool_goal_write(arguments: dict[str, Any]) -> str:
-    goal = normalize_goal_args(arguments)
-    status = "completed" if goal["done"] else f"{goal['progress']}%"
-    return f"Updated goal: {goal['title']} ({status})"
-
-
-def _tool_todo_write(arguments: dict[str, Any]) -> str:
-    todos = normalize_todo_items(arguments)
-    if not todos:
-        return "Error executing tool: todo_write requires a non-empty `todos` array"
-    lines = [
-        f"- [{t['status']}] {t['id']}: {t['content']}" for t in todos
-    ]
-    completed_n = sum(1 for t in todos if t.get("status") == "completed")
-    in_progress_n = sum(1 for t in todos if t.get("status") == "in_progress")
-    tip = ""
-    # Coach models that batch-complete everything after sitting on step 1.
-    if completed_n >= 3 and in_progress_n <= 1 and completed_n + in_progress_n >= len(todos) - 1:
-        tip = (
-            "\nNote for next turns: update statuses step-by-step "
-            "(complete current + set next in_progress) so Chat progress is visible — "
-            "avoid marking many items completed in a single todo_write."
-        )
-    return f"Updated {len(todos)} todo(s):\n" + "\n".join(lines) + tip
-
-
-def _tool_remember_preference(arguments: dict[str, Any]) -> str:
-    from src.cross_session_memory import add_entry
-    from src.preferences_storage import load_cross_session_memory_enabled
-
-    if not load_cross_session_memory_enabled():
-        return (
-            "Error executing tool: cross-session memory is disabled in Settings. "
-            "Ask the user to enable Memory first."
-        )
-    text = str(arguments.get("text") or "").strip()
-    if not text:
-        return "Error executing tool: remember_preference requires `text`"
-    run_id = _bg_job_run_id() or ""
-    try:
-        entry = add_entry(text, source_run_id=run_id or None)
-    except ValueError as exc:
-        return f"Error executing tool: {exc}"
-    return json.dumps({"ok": True, "id": entry["id"], "text": entry["text"]}, ensure_ascii=False)
-
-
-def _tool_submit_diff_summary(arguments: dict[str, Any]) -> str:
-    card = normalize_diff_summary(arguments, enrich=True)
-    if not card["files"]:
-        return "Error executing tool: submit_diff_summary requires a non-empty `files` array"
-    lines = [
-        f"- [{f.get('status') or 'M'}] {f['path']}"
-        + (f": {f['summary']}" if f.get("summary") else "")
-        for f in card["files"]
-    ]
-    summary = card.get("summary") or ""
-    summary_line = f"\nSummary: {summary}" if summary else ""
-    return (
-        f"Diff summary published: {card['title']}\n"
-        + "\n".join(lines)
-        + summary_line
-        + "\nUser can review readable diffs in Chat; continue or call submit_verification."
-    )
-
-
-def _tool_submit_verification(arguments: dict[str, Any]) -> str:
-    report = normalize_verification_report(arguments)
-    if not report["steps"]:
-        return "Error executing tool: submit_verification requires a non-empty `steps` array"
-    lines = [
-        f"- [{s['status']}] {s['name']}"
-        + (f": {s['detail']}" if s.get("detail") else "")
-        for s in report["steps"]
-    ]
-    actions = report.get("nextActions") or []
-    action_block = ""
-    if actions:
-        action_block = "\nNext actions:\n" + "\n".join(f"- {a}" for a in actions)
-    summary = report.get("summary") or ""
-    summary_line = f"\nSummary: {summary}" if summary else ""
-    return (
-        f"Verification {report['conclusion'].upper()}: {report['title']}\n"
-        + "\n".join(lines)
-        + summary_line
-        + action_block
-        + (
-            "\nDo not claim the task succeeded; address failed steps or next actions."
-            if report["conclusion"] == "failed"
-            else "\nChecks passed; you may summarize completion for the user."
-        )
-    )
-
-
-def _tool_ask_user_question(arguments: dict[str, Any]) -> str:
-    q = normalize_question_args(arguments)
-    selected = arguments.get("selected")
-    if isinstance(selected, dict):
-        label = str(selected.get("label") or "").strip()
-        oid = str(selected.get("id") or "").strip()
-        if label:
-            extra = f" (id={oid})" if oid else ""
-            return (
-                f"User selected: {label}{extra}. "
-                "Proceed with this choice; do not ask the same question again."
-            )
-    if isinstance(selected, str) and selected.strip():
-        return (
-            f"User selected: {selected.strip()}. "
-            "Proceed with this choice; do not ask the same question again."
-        )
-    opts = ", ".join(o["label"] for o in q["options"]) or "(none)"
-    return (
-        f"Question presented to user: {q['question']}\n"
-        f"Options: {opts}\n"
-        "Awaiting selection."
-    )
-
-
-def _tool_propose_plan(arguments: dict[str, Any]) -> str:
-    plan = normalize_plan_args(arguments)
-    steps = plan["steps"]
-    numbered = "\n".join(f"{i}. {step}" for i, step in enumerate(steps, 1)) or "(no steps)"
-    summary = plan["summary"]
-    extra = f"\nRationale: {summary}" if summary else ""
-    return (
-        f"Plan approved by the user: {plan['title']}\n"
-        f"Steps:\n{numbered}{extra}\n"
-        "Proceed to implement these steps with clutch-tools. "
-        "Do not call propose_plan again unless the user asks to revise."
-    )
-
-
-def _tool_apply_patch(arguments: dict[str, Any]) -> str:
-    from src.apply_patch import ApplyPatchError, apply_patch_in_workspace, extract_patch_paths, format_apply_patch_result
-    from src.artifact_layout import (
-        block_html_for_non_page_intent,
-        current_user_turn_text,
-        rewrite_apply_patch_paths,
-    )
-
-    patch = str(arguments.get("patch", "")).strip()
-    if not patch:
-        return "Error executing tool: apply_patch requires non-empty `patch`"
-    user_text = current_user_turn_text()
-    for path in extract_patch_paths(patch):
-        blocked = block_html_for_non_page_intent(path, user_text=user_text)
-        if blocked:
-            return f"Error executing tool: {blocked}"
-    patch, relocate_notes = rewrite_apply_patch_paths(patch, user_text=user_text)
-    run_id = _bg_job_run_id()
-    if run_id:
-        from src.file_rewind import snapshot_paths_before_write
-
-        snapshot_paths_before_write(run_id, extract_patch_paths(patch))
-    try:
-        result = format_apply_patch_result(apply_patch_in_workspace(patch))
-    except ApplyPatchError as exc:
-        return f"Error executing tool: {exc}"
-    if relocate_notes:
-        note = "; ".join(relocate_notes)
-        return f"{result}\n[Clutch] Relocated chat deliverable(s) under .clutch/artifacts/: {note}"
-    return result
-
-
-def _tool_generate_image(arguments: dict[str, Any]) -> str:
-    from src.image_router import (
-        format_image_reply,
-        generate_image_for_model,
-        persist_generated_image,
-        resolve_configured_image_model,
-    )
-
-    prompt = str(arguments.get("prompt") or "").strip()
-    if not prompt:
-        return "Error executing tool: generate_image requires `prompt`"
-    resolved = resolve_configured_image_model()
-    if resolved is None:
-        return (
-            "Error executing tool: no image model API key configured. "
-            "Add an image model key in Settings → Models (e.g. Agnes Image), "
-            "then retry generate_image. Do NOT write an HTML page as a substitute."
-        )
-    spec, api_key = resolved
-    stem = str(arguments.get("filename_stem") or "").strip() or None
-    try:
-        result = generate_image_for_model(spec, prompt, api_key=api_key)
-        result = persist_generated_image(result, filename_stem=stem)
-    except Exception as exc:
-        return (
-            f"Error executing tool: image generation failed ({exc}). "
-            "Do NOT write an HTML page as a substitute."
-        )
-    local = str(result.get("local_media_path") or "").strip()
-    # Keep tool payload small (no multi‑MB base64 in the ReAct transcript).
-    payload = {
-        "ok": True,
-        "model_id": spec.id,
-        "local_media_path": local,
-        "message": (
-            f"Image generated with {spec.name} and saved to `{local}`. "
-            "Include that path in your final reply; do not invent an HTML substitute."
-            if local
-            else f"Image generated with {spec.name}."
-        ),
-        # Optional short preview for Chat UI if the runner surfaces tool markdown later.
-        "preview_markdown": format_image_reply(result) if local else "",
-    }
-    return json.dumps(payload, ensure_ascii=False)
-
-
-def _tool_generate_video(arguments: dict[str, Any]) -> str:
-    from src.video_router import (
-        format_video_reply,
-        generate_video_for_model,
-        persist_generated_video,
-        resolve_configured_video_model,
-    )
-
-    prompt = str(arguments.get("prompt") or "").strip()
-    if not prompt:
-        return "Error executing tool: generate_video requires `prompt`"
-    resolved = resolve_configured_video_model()
-    if resolved is None:
-        return (
-            "Error executing tool: no video model API key configured. "
-            "Add a video model key in Settings → Models (e.g. Agnes Video), "
-            "then retry generate_video. Do NOT write an HTML page as a substitute."
-        )
-    spec, api_key = resolved
-    try:
-        result = generate_video_for_model(spec, prompt, api_key=api_key)
-        result = persist_generated_video(result)
-    except Exception as exc:
-        return (
-            f"Error executing tool: video generation failed ({exc}). "
-            "Do NOT write an HTML page as a substitute."
-        )
-    local = str(result.get("local_media_path") or "").strip()
-    payload = {
-        "ok": True,
-        "model_id": spec.id,
-        "local_media_path": local,
-        "message": (
-            f"Video generated with {spec.name} and saved to `{local}`."
-            if local
-            else f"Video generated with {spec.name}."
-        ),
-        "preview_markdown": format_video_reply(result) if local else "",
-    }
-    return json.dumps(payload, ensure_ascii=False)
-
-
-def _tool_read_file(arguments: dict[str, Any]) -> str:
-    from src.ignore_rules import ignored_path_message, is_ignored_path
-    from src.rich_read_util import (
-        is_rich_read_path,
-        read_image_workspace_file,
-        read_pdf_workspace_file,
-    )
-    from src.workspace import WorkspaceError, require_workspace, resolve_allowed_path
-
-    rel = str(arguments.get("path", "")).strip()
-    if not rel:
-        return "Error executing tool: read_file requires `path`"
-    try:
-        root = require_workspace()
-        if is_ignored_path(root, rel):
-            return f"Error executing tool: {ignored_path_message(rel)}"
-        target = resolve_allowed_path(rel)
-    except WorkspaceError as exc:
-        return f"Error executing tool: {exc}"
-    if not target.is_file():
-        return f"Error executing tool: not a file: {rel}"
-    if is_rich_read_path(target):
-        if target.suffix.lower() == ".pdf":
-            return read_pdf_workspace_file(target)
-        return read_image_workspace_file(target)
-    text = target.read_text(encoding="utf-8", errors="replace")
-    lines = text.splitlines()
-    offset = arguments.get("offset")
-    limit = arguments.get("limit")
-    start = 0
-    if offset is not None:
-        try:
-            start = max(0, int(offset) - 1)
-        except (TypeError, ValueError):
-            start = 0
-    end = len(lines)
-    if limit is not None:
-        try:
-            end = min(len(lines), start + max(0, int(limit)))
-        except (TypeError, ValueError):
-            pass
-    sliced = lines[start:end]
-    numbered = [f"{start + i + 1}|{line}" for i, line in enumerate(sliced)]
-    body = "\n".join(numbered)
-    if len(body) > _MAX_READ_CHARS:
-        body = body[:_MAX_READ_CHARS] + "\n…[truncated]"
-    return body or "(empty file)"
-
-
-def _tool_list_dir(arguments: dict[str, Any]) -> str:
-    from src.ignore_rules import is_ignored_path
-    from src.workspace import WorkspaceError, require_workspace, resolve_allowed_path, to_workspace_relative
-
-    rel = str(arguments.get("path") or ".").strip() or "."
-    try:
-        root = require_workspace()
-        if is_ignored_path(root, rel, is_dir=True):
-            from src.ignore_rules import ignored_path_message
-
-            return f"Error executing tool: {ignored_path_message(rel)}"
-        target = resolve_allowed_path(rel)
-    except WorkspaceError as exc:
-        return f"Error executing tool: {exc}"
-    if not target.is_dir():
-        return f"Error executing tool: not a directory: {rel}"
-    entries: list[str] = []
-    try:
-        children = sorted(target.iterdir(), key=lambda p: (not p.is_dir(), p.name.lower()))
-    except OSError as exc:
-        return f"Error executing tool: {exc}"
-    for child in children[:_MAX_LIST_ENTRIES]:
-        child_rel = to_workspace_relative(str(child)) or child.name
-        if is_ignored_path(root, child_rel, is_dir=child.is_dir()):
-            continue
-        suffix = "/" if child.is_dir() else ""
-        entries.append(f"{child.name}{suffix}")
-    extra = len(children) - _MAX_LIST_ENTRIES
-    if extra > 0:
-        entries.append(f"…and {extra} more")
-    return "\n".join(entries) if entries else "(empty directory)"
-
-
-def _tool_grep(arguments: dict[str, Any]) -> str:
-    from src.ignore_rules import is_ignored_path
-    from src.workspace import WorkspaceError, require_workspace, resolve_allowed_path, to_workspace_relative
-
-    pattern = str(arguments.get("pattern", ""))
-    if not pattern:
-        return "Error executing tool: grep requires `pattern`"
-    scope = str(arguments.get("path") or ".").strip() or "."
-    case_insensitive = bool(arguments.get("case_insensitive"))
-    try:
-        root = require_workspace()
-        scope_path = resolve_allowed_path(scope)
-        if is_ignored_path(root, scope, is_dir=scope_path.is_dir()):
-            from src.ignore_rules import ignored_path_message
-
-            return f"Error executing tool: {ignored_path_message(scope)}"
-    except WorkspaceError as exc:
-        return f"Error executing tool: {exc}"
-
-    rg = shutil.which("rg")
-    if rg:
-        cmd = [rg, "--line-number", "--no-heading", "--color", "never", "-m", str(_MAX_GREP_HITS)]
-        if case_insensitive:
-            cmd.append("-i")
-        cmd.extend(["--", pattern, str(scope_path)])
-        try:
-            proc = subprocess.run(
-                cmd,
-                cwd=str(root),
-                capture_output=True,
-                text=True,
-                timeout=30,
-                check=False,
-            )
-        except subprocess.TimeoutExpired:
-            return "Error executing tool: grep timed out"
-        if proc.returncode not in (0, 1):
-            err = (proc.stderr or proc.stdout or "rg failed").strip()
-            return f"Error executing tool: {err[:500]}"
-        out = (proc.stdout or "").strip()
-        if out:
-            filtered: list[str] = []
-            for line in out.splitlines():
-                rel = line.split(":", 1)[0] if ":" in line else ""
-                if rel and is_ignored_path(root, rel):
-                    continue
-                filtered.append(line)
-            out = "\n".join(filtered)
-        return out or "(no matches)"
-
-    flags = re.IGNORECASE if case_insensitive else 0
-    try:
-        regex = re.compile(pattern, flags)
-    except re.error as exc:
-        return f"Error executing tool: invalid pattern: {exc}"
-    hits: list[str] = []
-    paths = [scope_path] if scope_path.is_file() else scope_path.rglob("*")
-    for path in paths:
-        if not path.is_file():
-            continue
-        try:
-            rel = str(path.relative_to(root))
-        except ValueError:
-            rel = str(path)
-        if is_ignored_path(root, rel):
-            continue
-        try:
-            if path.stat().st_size > 2_000_000:
-                continue
-            text = path.read_text(encoding="utf-8", errors="replace")
-        except OSError:
-            continue
-        for idx, line in enumerate(text.splitlines(), start=1):
-            if regex.search(line):
-                hits.append(f"{rel}:{idx}:{line}")
-                if len(hits) >= _MAX_GREP_HITS:
-                    return "\n".join(hits)
-    return "\n".join(hits) if hits else "(no matches)"
-
-
-def _tool_search_replace(arguments: dict[str, Any]) -> str:
-    from src.workspace import WorkspaceError, resolve_allowed_path, to_workspace_relative
-
-    rel = str(arguments.get("path", "")).strip()
-    old = arguments.get("old_string")
-    new = arguments.get("new_string")
-    replace_all = bool(arguments.get("replace_all"))
-    if not rel:
-        return "Error executing tool: search_replace requires `path`"
-    if old is None or new is None:
-        return "Error executing tool: search_replace requires `old_string` and `new_string`"
-    old_s = str(old)
-    new_s = str(new)
-    if old_s == new_s:
-        return "Error executing tool: old_string and new_string are identical"
-    if not old_s:
-        return "Error executing tool: old_string must be non-empty"
-    try:
-        target = resolve_allowed_path(rel)
-    except WorkspaceError as exc:
-        return f"Error executing tool: {exc}"
-    if not target.is_file():
-        return f"Error executing tool: not a file: {rel}"
-    run_id = _bg_job_run_id()
-    if run_id:
-        from src.file_rewind import snapshot_before_write
-
-        snapshot_before_write(run_id, rel)
-    text = target.read_text(encoding="utf-8", errors="replace")
-    count = text.count(old_s)
-    if count == 0:
-        return "Error executing tool: old_string not found in file"
-    if count > 1 and not replace_all:
-        return (
-            f"Error executing tool: old_string found {count} times; "
-            "pass replace_all=true or provide a more unique old_string"
-        )
-    updated = text.replace(old_s, new_s) if replace_all else text.replace(old_s, new_s, 1)
-    target.write_text(updated, encoding="utf-8")
-    rel_out = to_workspace_relative(str(target)) or rel
-    replaced = count if replace_all else 1
-    return json.dumps(
-        {"ok": True, "path": rel_out, "replacements": replaced, "changed_paths": [rel_out]},
-        ensure_ascii=False,
-    )
-
-
-def _bg_job_run_id() -> str | None:
-    from src.bg_jobs import get_bg_job_context
-
-    ctx = get_bg_job_context()
-    if not ctx:
-        return None
-    run_id = str(ctx.get("run_id") or "").strip()
-    return run_id or None
-
-
-def _tool_run_terminal_cmd(arguments: dict[str, Any]) -> str:
-    from src.workspace import WorkspaceError, assert_strict_sandbox_command, require_workspace
-
-    command = str(arguments.get("command", "")).strip()
-    if not command:
-        return "Error executing tool: run_terminal_cmd requires `command`"
-    try:
-        root = require_workspace()
-        assert_strict_sandbox_command(command, root)
-    except WorkspaceError as exc:
-        return f"Error executing tool: {exc}"
-
-    if bool(arguments.get("background")):
-        run_id = _bg_job_run_id()
-        if not run_id:
-            return (
-                "Error executing tool: background commands require an active Chat run context"
-            )
-        from src.bg_jobs import start_job
-
-        try:
-            job = start_job(run_id, command, str(root))
-        except ValueError as exc:
-            return f"Error executing tool: {exc}"
-        return json.dumps(
-            {
-                "ok": True,
-                "job_id": job["id"],
-                "status": job["status"],
-                "title": job.get("title") or command[:60],
-            },
-            ensure_ascii=False,
-        )
-
-    try:
-        timeout = int(arguments.get("timeout_sec") or _DEFAULT_CMD_TIMEOUT_S)
-    except (TypeError, ValueError):
-        timeout = _DEFAULT_CMD_TIMEOUT_S
-    timeout = max(1, min(timeout, 300))
-
-    run_id = _bg_job_run_id()
-    if run_id:
-        from src.foreground_shell import start_foreground, wait_foreground
-
-        try:
-            start_foreground(run_id, command, str(root))
-        except Exception as exc:
-            return f"Error executing tool: {exc}"
-        output, transferred, exit_code = wait_foreground(run_id, timeout_sec=float(timeout))
-        if transferred:
-            from src.bg_jobs import list_jobs
-
-            jobs = list_jobs(run_id)
-            job_id = jobs[-1]["id"] if jobs else ""
-            return json.dumps(
-                {
-                    "ok": True,
-                    "transferred_to_background": True,
-                    "job_id": job_id,
-                    "status": "running",
-                    "title": command[:60],
-                    "output_prefix": output[:2000] if output else "",
-                },
-                ensure_ascii=False,
-            )
-        if exit_code is None:
-            return f"Error executing tool: command timed out after {timeout}s"
-        header = f"exit_code={exit_code}\n"
-        combined = output
-        if len(combined) > _MAX_CMD_OUTPUT_CHARS:
-            combined = combined[:_MAX_CMD_OUTPUT_CHARS] + "\n…[truncated]"
-        return header + (combined if combined.strip() else "(no output)")
-
-    shell = os.environ.get("SHELL") or ("cmd.exe" if os.name == "nt" else "/bin/bash")
-    try:
-        proc = subprocess.run(
-            command,
-            shell=True,
-            cwd=str(root),
-            capture_output=True,
-            text=True,
-            timeout=timeout,
-            check=False,
-            env={**os.environ, "PWD": str(root)},
-            executable=shell if os.name != "nt" and Path(shell).is_file() else None,
-        )
-    except subprocess.TimeoutExpired:
-        return f"Error executing tool: command timed out after {timeout}s"
-    stdout = proc.stdout or ""
-    stderr = proc.stderr or ""
-    combined = stdout
-    if stderr.strip():
-        combined = f"{stdout}\n[stderr]\n{stderr}" if stdout else f"[stderr]\n{stderr}"
-    if len(combined) > _MAX_CMD_OUTPUT_CHARS:
-        combined = combined[:_MAX_CMD_OUTPUT_CHARS] + "\n…[truncated]"
-    header = f"exit_code={proc.returncode}\n"
-    return header + (combined if combined.strip() else "(no output)")
-
-
-def _tool_list_background_jobs(arguments: dict[str, Any]) -> str:
-    del arguments
-    run_id = _bg_job_run_id()
-    if not run_id:
-        return "Error executing tool: list_background_jobs requires an active Chat run context"
-    from src.bg_jobs import list_jobs
-
-    return json.dumps(list_jobs(run_id), ensure_ascii=False)
-
-
-def _tool_kill_background_job(arguments: dict[str, Any]) -> str:
-    run_id = _bg_job_run_id()
-    if not run_id:
-        return "Error executing tool: kill_background_job requires an active Chat run context"
-    job_id = str(arguments.get("job_id") or "").strip()
-    if not job_id:
-        return "Error executing tool: kill_background_job requires `job_id`"
-    from src.bg_jobs import kill_job
-
-    killed = kill_job(run_id, job_id)
-    if killed is None:
-        return f"Error executing tool: background job `{job_id}` not found"
-    return json.dumps(killed, ensure_ascii=False)
-
-
-def _run_git(args: list[str], *, cwd: Path) -> subprocess.CompletedProcess[str]:
-    return subprocess.run(
-        ["git", *args],
-        cwd=str(cwd),
-        capture_output=True,
-        text=True,
-        check=False,
-        timeout=60,
-    )
-
-
-def _tool_git_status(arguments: dict[str, Any]) -> str:
-    del arguments
-    from src.workspace import WorkspaceError, require_workspace
-
-    try:
-        root = require_workspace()
-    except WorkspaceError as exc:
-        return f"Error executing tool: {exc}"
-    proc = _run_git(["status", "--short", "--branch"], cwd=root)
-    if proc.returncode != 0:
-        err = (proc.stderr or proc.stdout or "git status failed").strip()
-        return f"Error executing tool: {err}"
-    out = (proc.stdout or "").strip()
-    return out or "(clean)"
-
-
-def _tool_git_diff(arguments: dict[str, Any]) -> str:
-    from src.workspace import WorkspaceError, require_workspace
-
-    try:
-        root = require_workspace()
-    except WorkspaceError as exc:
-        return f"Error executing tool: {exc}"
-    staged = bool(arguments.get("staged"))
-    paths = [
-        str(p).strip()
-        for p in (arguments.get("paths") or [])
-        if str(p).strip()
-    ]
-    args = ["diff", "--staged"] if staged else ["diff"]
-    if paths:
-        args.extend(["--", *paths])
-    proc = _run_git(args, cwd=root)
-    if proc.returncode not in (0, 1):
-        err = (proc.stderr or proc.stdout or "git diff failed").strip()
-        return f"Error executing tool: {err}"
-    out = proc.stdout or ""
-    if len(out) > _MAX_CMD_OUTPUT_CHARS:
-        out = out[:_MAX_CMD_OUTPUT_CHARS] + "\n…[truncated]"
-    return out.strip() or "(no diff)"
-
-
-def _tool_git_commit(arguments: dict[str, Any]) -> str:
-    from src.workspace import WorkspaceError, require_workspace
-
-    message = str(arguments.get("message") or "").strip()
-    if not message:
-        return "Error executing tool: git_commit requires `message`"
-    try:
-        root = require_workspace()
-    except WorkspaceError as exc:
-        return f"Error executing tool: {exc}"
-    paths = [
-        str(p).strip()
-        for p in (arguments.get("paths") or [])
-        if str(p).strip()
-    ]
-    if paths:
-        add = _run_git(["add", "--", *paths], cwd=root)
-    else:
-        add = _run_git(["add", "-A"], cwd=root)
-    if add.returncode != 0:
-        err = (add.stderr or add.stdout or "git add failed").strip()
-        return f"Error executing tool: {err}"
-    commit = _run_git(["commit", "-m", message], cwd=root)
-    if commit.returncode != 0:
-        err = (commit.stderr or commit.stdout or "git commit failed").strip()
-        return f"Error executing tool: {err}"
-    head = _run_git(["rev-parse", "--short", "HEAD"], cwd=root)
-    sha = (head.stdout or "").strip() if head.returncode == 0 else ""
-    return json.dumps(
-        {
-            "ok": True,
-            "message": message,
-            "sha": sha,
-            "stdout": (commit.stdout or "").strip(),
-        },
-        ensure_ascii=False,
-    )
-
-
-def _tool_web_fetch(arguments: dict[str, Any]) -> str:
-    from src.web_fetch_util import (
-        extract_serp_query,
-        fetch_url_text,
-        is_search_engine_serp_url,
-        serp_redirect_error_message,
-    )
-
-    url = str(arguments.get("url") or "").strip()
-    if not url:
-        return "Error executing tool: web_fetch requires `url`"
-    try:
-        timeout = int(arguments.get("timeout_sec") or 20)
-    except (TypeError, ValueError):
-        timeout = 20
-
-    # Flash models often web_fetch google.com/search — rewrite to web_search so
-    # the turn can continue (and loop fuse is not burned on policy rejects).
-    if is_search_engine_serp_url(url):
-        from src.preferences_storage import load_allow_network
-        from src.web_search_util import search_web
-
-        query = extract_serp_query(url)
-        if load_allow_network() and query:
-            try:
-                payload = search_web(query, max_results=5)
-            except Exception as exc:
-                return f"Error executing tool: {serp_redirect_error_message(url)} ({exc})"
-            payload = {
-                **payload,
-                "redirected_from_web_fetch": True,
-                "original_url": url,
-                "note": (
-                    "You called web_fetch on a search-engine results URL. "
-                    "Clutch ran web_search instead. Next: web_fetch at most 1–2 "
-                    "concrete article URLs from results[], then answer or write the HTML — "
-                    "do not fetch google.com/search / bing.com/search again."
-                ),
-            }
-            return json.dumps(payload, ensure_ascii=False)
-        return f"Error executing tool: {serp_redirect_error_message(url)}"
-
-    try:
-        payload = fetch_url_text(url, timeout_sec=timeout)
-    except ValueError as exc:
-        return f"Error executing tool: {exc}"
-    return json.dumps(payload, ensure_ascii=False)
-
-
-def _tool_web_search(arguments: dict[str, Any]) -> str:
-    from src.preferences_storage import load_allow_network
-    from src.web_search_util import search_web
-
-    if not load_allow_network():
-        return (
-            "Error executing tool: web_search is disabled. "
-            "Enable Settings → General → Allow network to search the web."
-        )
-    query = str(arguments.get("query") or "").strip()
-    if not query:
-        return "Error executing tool: web_search requires `query`"
-    try:
-        max_results = int(arguments.get("max_results") or 5)
-    except (TypeError, ValueError):
-        max_results = 5
-    try:
-        payload = search_web(query, max_results=max_results)
-    except ValueError as exc:
-        return f"Error executing tool: {exc}"
-    except Exception as exc:
-        from src.web_fetch_util import _friendly_network_error
-
-        return f"Error executing tool: web search failed: {_friendly_network_error(exc)}"
-    return json.dumps(payload, ensure_ascii=False)
-
-
-def _tool_diagnostics(arguments: dict[str, Any]) -> str:
-    import json
-
-    from src.code_diagnostics import run_code_diagnostics, store_pending_diagnostics
-    from src.workspace import WorkspaceError, require_workspace
-
-    try:
-        root = require_workspace()
-    except WorkspaceError as exc:
-        return f"Error executing tool: {exc}"
-    paths_arg = arguments.get("paths")
-    paths: list[str] | None = None
-    if isinstance(paths_arg, list):
-        paths = [str(item) for item in paths_arg if str(item).strip()]
-    issues = run_code_diagnostics(root, paths)
-    run_id = _bg_job_run_id()
-    if run_id:
-        store_pending_diagnostics(run_id, issues)
-    return json.dumps({"count": len(issues), "issues": issues}, ensure_ascii=False)
